+++ packages/waggle/src/communities/manager.ts
@@ -69,6 +69,9 @@ export default class CommunitiesManager {
   public initStorage = async (peerId: PeerId, onionAddress: string, port: number, bootstrapMultiaddrs: string[], certs: CertsData): Promise<string> => {
     const listenAddrs = `/dns4/${onionAddress}/tcp/${port}/wss`
     const peerIdB58string = peerId.toB58String()
+    if (bootstrapMultiaddrs.length === 0) {
+      bootstrapMultiaddrs = [listenAddrs]
+    }
     const libp2pObj = await this.connectionsManager.initLibp2p(peerId, listenAddrs, bootstrapMultiaddrs, certs)
     const storage = new this.connectionsManager.StorageCls(
       this.connectionsManager.zbayDir,