+++ packages/waggle/src/libp2p/connectionsManager.test.ts
@@ -72,7 +72,7 @@ describe('Connections manager', () => {
     })
     await connectionsManager.init()
     const network = await connectionsManager.createNetwork()
-    expect(network.hiddenService.onionAddress).toHaveLength(56)
+    expect(network.hiddenService.onionAddress.split('.')[0]).toHaveLength(56)
     expect(network.hiddenService.privateKey).toHaveLength(99)
     const peerId = await PeerId.createFromJSON(network.peerId)
     expect(PeerId.isPeerId(peerId)).toBeTruthy()