+++ packages/waggle/src/libp2p/tests/client-server.ts
@@ -1,4 +1,4 @@
-import { Time, getCrypto, Certificate } from 'pkijs'
+import { Time, Certificate } from 'pkijs'
 
 import { createUserCert, createUserCsr, createRootCA, configCrypto } from '@zbayapp/identity'
 import { RootCA } from '@zbayapp/identity/lib/generateRootCA'
@@ -37,7 +37,7 @@ function formatPEM(pemString: string) {
   return resultString
 }
 
-export const createUsersCerts = async (onion: string, rootCert: RootCA): Promise<{ userCert: Buffer, userKey: Buffer }> => {
+export const createUsersCerts = async (onion: string, rootCert: RootCA): Promise<{ userCert: string, userKey: string }> => {
   const userData = {
     zbayNickname: 'dev99damian1',
     commonName: onion,
@@ -54,8 +54,8 @@ export const createUsersCerts = async (onion: string, rootCert: RootCA): Promise
   const userCert = await createUserCert(rootCert.rootCertString, rootCert.rootKeyString, user.userCsr, notBeforeDate, notAfterDate)
 
   return {
-    userCert: dumpPEM('CERTIFICATE', userCert.userCertObject.certificate.toSchema(true).toBER(false)),
-    userKey: dumpPEM('PRIVATE KEY', await getCrypto().exportKey('pkcs8', user.pkcs10.privateKey))
+    userCert: userCert.userCertString,
+    userKey: user.userKey
   }
 }
 
@@ -68,8 +68,8 @@ export const createCertificatesTestHelper = async (onion1, onion2) => {
   const userData2 = await createUsersCerts(onion2, rootCert)
 
   const pems = {
-    ca: dumpPEM('CERTIFICATE', rootCert.rootObject.certificate.toSchema(true).toBER(false)),
-    ca_key: dumpPEM('PRIVATE KEY', await getCrypto().exportKey('pkcs8', rootCert.rootObject.privateKey)),
+    ca: rootCert.rootCertString,
+    ca_key: rootCert.rootKeyString,
 
     servCert: userData1.userCert,
     servKey: userData1.userKey,
@@ -77,6 +77,5 @@ export const createCertificatesTestHelper = async (onion1, onion2) => {
     userCert: userData2.userCert,
     userKey: userData2.userKey
   }
-
   return pems
 }