+++ packages/waggle/src/registration/registrationService.test.ts
@@ -2,7 +2,7 @@ import { createRootCA, createUserCert, createUserCsr, verifyUserCert, configCryp
 import { SocksProxyAgent } from 'socks-proxy-agent'
 import { CertificateRegistration } from '.'
 import { Time } from 'pkijs'
-import { createLibp2p, createTmpDir, spawnTorProcess, TmpDir, tmpZbayDirPath, TorMock } from '../testUtils'
+import { createLibp2p, createTmpDir, spawnTorProcess, TmpDir, tmpZbayDirPath, TorMock, dataFromRootPems } from '../testUtils'
 import { DummyIOServer, getPorts, Ports } from '../utils'
 import fetch, { Response } from 'node-fetch'
 import { Tor } from '../torManager'
@@ -10,6 +10,7 @@ import { RootCA } from '@zbayapp/identity/lib/generateRootCA'
 import { Storage } from '../storage'
 import PeerId from 'peer-id'
 import { DataFromPems } from '../common/types'
+// import {registerOwnerCertificate} from './index'
 jest.setTimeout(50_000)
 
 async function registerUserTest(csr: string, socksPort: number, localhost: boolean = true): Promise<Response> {
@@ -146,6 +147,7 @@ describe('Registration service', () => {
     const isProperUserCert = await verifyUserCert(certRoot.rootCertString, responseData.certificate)
     expect(isProperUserCert.result).toBe(true)
     expect(responseData.peers.length).toBe(1)
+    expect(responseData.rootCa).toBe(certRoot.rootCertString)
   })
 
   it('returns 403 if username already exists', async () => {
@@ -208,5 +210,18 @@ describe('Registration service', () => {
     const csr = 'MIIBFTCBvAIBADAqMSgwFgYKKwYBBAGDjBsCARMIdGVzdE5hbWUwDgYDVQQDEwdaYmF5IENBMFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEGPGHpJzE/CvL7l/OmTSfYQrhhnWQrYw3GgWB1raCTSeFI/MDVztkBOlxwdUWSm10+1OtKVUWeMKaMtyIYFcPPqAwMC4GCSqGSIb3DQEJDjEhMB8wHQYDVR0OBBYEFLjaEh+cnNhsi5qDsiMB/ZTzZFfqMAoGCCqGSM49BAMCA0gAMEUCIFwlob/Igab05EozU0e/lsG7c9BxEy4M4c4Jzru2vasGAiEAqFTQuQr/mVqTHO5vybWm/iNDk8vh88K6aBCCGYqIfdw='
     const response = await registerUserTest(csr, ports.socksPort)
     expect(response.status).toEqual(400)
+  })
+  it('registers owner certificate', async () => {
+    const csr = await createUserCsr({
+      zbayNickname: 'userName',
+      commonName: 'nqnw4kc4c77fb47lk52m5l57h4tcxceo7ymxekfn7yh5m66t4jv2olad.onion',
+      peerId: 'Qmf3ySkYqLET9xtAtDzvAr5Pp3egK1H3C5iJAZm1SpLEp6',
+      dmPublicKey: 'testdmPublicKey',
+      signAlg: configCrypto.signAlg,
+      hashAlg: configCrypto.hashAlg
+    })
+    const certificate = await CertificateRegistration.registerOwnerCertificate(csr.userCsr, dataFromRootPems)
+    const isProperUserCert = await verifyUserCert(dataFromRootPems.certificate, certificate)
+    expect(isProperUserCert.result).toBe(true)
   })
 })